[#88] feat: 초대하기 기능 사용 시 초대 내역 컴포넌트 업데이트 구현
