+++ src/pages/[id]/mydashboard.tsx
@@ -1,11 +1,6 @@
-import { postDashboardsInvitations } from '@/api/dashboard';
 import DeleteDashBoardButton from '@/components/button/dashBoard/delete/deleteDashBoardButton';
 import ReturnButton from '@/components/button/dashBoard/return/returnButton';
 import DashboardHeader from '@/components/header/dashboardHeader/dashboardHeader';
-import ModalButtonSet from '@/components/modal/button/modalButtonSet';
-import DefaultInput from '@/components/modal/input/defaultInput/defaultInput';
-import InputLayout from '@/components/modal/input/inputLayout';
-import InputModal from '@/components/modal/inputModal/inputModal';
 import SideMenu from '@/components/sideMenu/SideMenu';
 import EditDashboard from '@/components/table/editDashboard/editDashboard';
 import InvitationList from '@/components/table/invitation/invitationList';
@@ -13,9 +8,6 @@ import MemberList from '@/components/table/member/memberList';
 import S from '@/pages/[id]/mydashboard.module.css';
 import { GetServerSidePropsContext } from 'next';
 import { useEffect, useState } from 'react';
-import { FieldValues, useForm } from 'react-hook-form';
-import CommonStyle from '@/components/modal/modalCommon.module.css';
-import { useMutation } from '@tanstack/react-query';
 
 export async function getServerSideProps(context: GetServerSidePropsContext) {
   if (!context.params) {
@@ -42,33 +34,12 @@ function DashboardEditPage({ dashboardId }: DashboardEditPageProps) {
   const [flag, setFlag] = useState(false);
   const [invitationFlag, setInvitationFlag] = useState(false);
   const [memberFlag, setMemberFlag] = useState(false);
-  const mutation = useMutation({
-    mutationFn: (data: FieldValues) =>
-      postDashboardsInvitations(dashboardId, data),
-    onError: (error) => {
-      alert(error);
-    },
-  });
+
   const handleModal = () => {
     setIsModalOpen(!isModalOpen);
   };
 
-  const methods = useForm<FieldValues>({
-    mode: 'onChange',
-    defaultValues: {
-      email: '',
-    },
-  });
-
-  const { handleSubmit, control, reset } = methods;
-
-  function handleAddTodo(data: FieldValues) {
-    mutation.mutate(data);
-    setIsModalOpen(false);
-    reset();
-  }
-
-  // 모달이 열릴 경우 백그라운드 스크롤 방지
+  // // 모달이 열릴 경우 백그라운드 스크롤 방지
   useEffect(() => {
     if (isModalOpen) {
       document.body.style.overflowY = 'hidden';
@@ -78,62 +49,42 @@ function DashboardEditPage({ dashboardId }: DashboardEditPageProps) {
   }, [isModalOpen]);
 
   return (
-    <>
-      <div className={S.mainOuter}>
-        <SideMenu pageId={Number(dashboardId)} flag={flag} />
-        <div className={S.main}>
-          <DashboardHeader
-            folder="1"
-            users={[]}
-            user={{
-              letter: '1',
-              name: 'kim',
-              color: 'yellow',
-              ownerFolder: { folder: '1' },
-            }}
-          />
-          <div className={S.mainContainer}>
-            <ReturnButton url={`/${dashboardId}`} />
-            <div className={S.tableContainer}>
-              <EditDashboard dashboardId={dashboardId} setFlag={setFlag} />
-              <MemberList
-                dashboardId={dashboardId}
-                memberFlag={memberFlag}
-                setMemberFlag={setMemberFlag}
-              />
-              <InvitationList
-                dashboardId={dashboardId}
-                onClick={handleModal}
-                invitationFlag={invitationFlag}
-                setInvitationFlag={setInvitationFlag}
-              />
-              <div className={S.marginDiv}></div>
-              <DeleteDashBoardButton />
-            </div>
+    <div className={S.mainOuter}>
+      <SideMenu pageId={Number(dashboardId)} flag={flag} />
+      <div className={S.main}>
+        <DashboardHeader
+          folder="1"
+          users={[]}
+          user={{
+            letter: '1',
+            name: 'kim',
+            color: 'yellow',
+            ownerFolder: { folder: '1' },
+          }}
+        />
+        <div className={S.mainContainer}>
+          <ReturnButton url={`/${dashboardId}`} />
+          <div className={S.tableContainer}>
+            <EditDashboard dashboardId={dashboardId} setFlag={setFlag} />
+            <MemberList
+              dashboardId={dashboardId}
+              memberFlag={memberFlag}
+              setMemberFlag={setMemberFlag}
+            />
+            <InvitationList
+              dashboardId={dashboardId}
+              onClick={handleModal}
+              invitationFlag={invitationFlag}
+              setInvitationFlag={setInvitationFlag}
+              isModalOpen={isModalOpen}
+              setIsModalOpen={setIsModalOpen}
+            />
+            <div className={S.marginDiv}></div>
+            <DeleteDashBoardButton />
           </div>
         </div>
       </div>
-      {isModalOpen && (
-        <InputModal onClick={handleModal} title={'초대하기'}>
-          <InputLayout label="이메일" isNecessary={false}>
-            <form
-              onSubmit={handleSubmit(handleAddTodo)}
-              className={CommonStyle.form}>
-              <DefaultInput
-                placeholder="이메일을 입력해 주세요"
-                control={control}
-                name="email"
-              />
-              <ModalButtonSet
-                isDelete={false}
-                submitButtonTitle="초대"
-                onClickCancel={handleModal}
-              />
-            </form>
-          </InputLayout>
-        </InputModal>
-      )}
-    </>
+    </div>
   );
 }
